Fix floating clone positions
Floating windows tend not to trigger size-changed at all. Make sure we position
the clone correctly at least once.

Might of course be better to just run allocateClone when the actor is allocated
still, but I'm not entirely sure if the frame and actor is necessarily synced at
that point.