kludges: refactor override system
Not in full use yet.

But the main point is saving the original method on init, supplying an optional
override which will take effect on enable. We reset all properties to their
original form on disable.

NOTE: a property is only ever saved the first time, but new overrides can be
supplied as much as you want.