action dispatcher: Introduce global pseudo focus while navigating
The action dispatcher would only allow actions to operate on tiled windows.

However some actions, like cycleWindowWidth, are able to operate on eg. scratch
windows, while it at the same time wants context between repeated cycles.

To solve this we add an explicitly tracked pseduo focus to the dispatcher.
Actions can now return a MetaWindow which then becomes the new focus.

This should pave the way for us to integrate eg. alt-tab with the dispatcher.