Implement window drag and drop
It's now possible to rearrange windows with the mouse:

Dragging the window up, or down, will scale its size gradually down.
When below a certain threshold it will pop out of the tiling leaving the
user free to attach it at any tiling edge.

If not dropped onto a tiling edge it will end up as a scratch/float window.