Do not animate the top bar
Hiding the top bar doesn't change the workarea, at least in 3.34.
Having scale_y = 0 does change the workarea. In other words animating
the top bar means the workarea doesn't change immediately.

This can cause trouble when going from fullscreen to maximized, as the
workarea won't be updated early enough, giving the maximized window the
full monitor.

The animation isn't really that useful, so removing it is a viable fix.

In the case of user hidden top bar we still need to update the workare,
so we update the scale directly.

Also make the hideTopbar logic somewhat readable.

closes #220