Bugfix: (X11 only?) fix occationally initial jump when moving window
The `global.get_pointer` is not necessarily in sync with the clutter motion event.

We use `global.get_pointer` to set the anchor in `begin`, but sometimes (fast
motions more likely to trigger) it's an extra event in the `motion` event queue.
The motion handler will in effect see a reality that occured before the anchor
was set, triggering a "jump".

Also notice that `global.get_pointer` seems to be ahead of the actual grab event
:/ meaning we get less responsive window moving. We don't have access the actual
click coordinates which triggered the move :(