fix: cannot switch windows using navigator on X11
When the keyboard dispatcher is dismissed before destroying the navigator, a focus signal is emitted to the starting window. This causes a ensureViewport, that changes the selected window to the first one, overriding the selected one.
Since the logic that activate the new window is done inside the destroy function, the selected window is wrong after the dismissDispatcher.

Not sure why this happens only on X11 and on the last GNOME version. I don't understand why the focused window is focused a second time, like if the navigator is stealing the focus.