Fix rounding errors in resizeWInc() and similar
Some applications change their window size to certain boundries, such as
terminal emulators to fit a whole number of text columns. When resizing
such windows with resizeWInc(), the window should grow in width to the
next 10% of the total width. However, if the terminal emulator reduces
that 10% to 9.9% to hit a whole number of text columns, the next call to
resizeWInc() will just try to increase that last 0.1% and the terminal
emulator will immediately take it back to hit a whole number of columns.

This patch changes the behaviour of resizeWInd() and friends to increase
the width with something between 5% and 15%, whatever hits a tenth of
the full width, instead of something between 0% and 10%.