Implemented proposal from #766
This approach calls Utils.getPointerCoords() (the unified touch broker)
when the event starts ONLY. Since the touch tracker fails to update once
the event is "adopted" by a widget, we still need to update the global
cursor's coordinates.

This superficially fixes #764, but REVERT this commit before working on
cleaning up touch handling!