Add EDGE focus mode
This mode is especially aimed at heavy touchpad swipe gesture users.

With the keyboard shortcuts and default focus mode, we already mostly
just get one edge of a window always snapped to an edge of the screen.

With swipe gestures however it is way too easy to end up at some
arbitrary scroll position, as the default snap behavior (based only on
windows getting of bounds) is not enough to ensure snapped edges.