Fail gracefully instead of crashing if grab fails
I have been using PaperWM on a daily basis for the past few months and it is great. One of the
problems that I kept running into repeatedly was https://github.com/paperwm/PaperWM/issues/991. This
issue was reported on Gnome 47. I ran into this problem on Gnome 44.

I attempted a naive fix which simply tries to clean-up all the state instead of crashing. This works
well; I have been using a version of this fix, applied on top of the gnome-44 branch, for the past
month and I have not suffered any `Could not grab modal` crashes.

It would be great if I could get some guidance about whether this fix looks logical, and whether
there might be a way to reproduce this code path somehow. We need to somehow get into
`preview_navigate` (which happens when one uses a keyboard shortcut to switch windows, I believe?)
but get there in a state where the keyboard is already grabbed by something else, thus causing
PaperWM to fail to grab the keyboard. I would really like to fix this issue within PaperWM because I
have not run into any other bug at all and this is a great project!

Signed-off-by: Siddharth Kannan <mail@siddharthkannan.in>