Focus the closest *visible* window
In the diagrams below, assume parens are workspaces, brackets are
windows currently showing on a monitor, and letters are windows.
Uppercase means focused, and assume all windows are 50% sized.

If we start like this:

    ([A b] c d) ([e f])

Then I use the mouse to move to the monitor to the right:

    ([a b] c d) ([E, f])

Then I use this keybind to move left again. I'd get whiplash as the
monitor on the left spins to move windows c and d back onscreen.

    (a b [c D]) ([u, f])

Now, instead it'll just pick the closest window on screen:

    ([a B] c d) ([e, f])

Which makes sense visually, since in this situation I'm thinking "I want
to move left" and what I see is this transition:

    [a b][E f]
    [a B][e f]

It might make sense to add some margin here in case window c is just
barely peeking a few pixels on screen (in which case we might still
prefer to focus window b).

Also, some small code cleanup: using more `const`

It's hard for me to keep track of what variables mean, when we are
re-assigning them. So, create a new variable to track the new space, and
make everything const so we know it's not being updated.

_Maybe_ re-using a variable saves a few bytes of memory, or a little bit
of extra garbage collection time, but this function is only being run
once per user input, so I doubt it's a significant performance impact.